backend/requirements.txt
asgiref==3.9.1
Django==4.2
django-cors-headers==4.0.0
django-filter==25.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
setuptools==80.9.0
sqlparse==0.5.3
